refact: update api exception to include error object
