+++ openai-client/src/commonTest/kotlin/com/aallam/openai/client/TestException.kt
@@ -1,0 +1,30 @@
+package com.aallam.openai.client
+
+import com.aallam.openai.api.exception.OpenAIAPIException
+import com.aallam.openai.api.model.ModelId
+import kotlinx.coroutines.test.runTest
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
+
+class TestException {
+
+    @Test
+    fun apiError() = runTest {
+        val openAI = OpenAI("sk-123")
+        val model = runCatching {
+            openAI.model(ModelId("davinci"))
+        }
+        assertTrue(model.isFailure)
+        val exception = model.exceptionOrNull()?.cause as OpenAIAPIException
+        assertEquals(401, exception.statusCode)
+        assertEquals(
+            "invalid_request_error",
+            exception.error.detail?.type
+        )
+        assertEquals(
+            "invalid_api_key",
+            exception.error.detail?.code
+        )
+    }
+}
