--- openai-core/src/commonMain/kotlin/com.aallam.openai.api/exception/OpenAIAPIError.kt
@@ -1,11 +1,0 @@
-package com.aallam.openai.api.exception
-
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-public data class OpenAIAPIError(
-    @SerialName("error")
-    val error: Error?
-)
